// src/lib/types.ts
export interface User {
  id: string;
  role: "client" | "mentor" | "admin";
  name: string;
  email: string;
  avatar: string;
  mentorId?: string;
  timezone?: string;
}